Set region to ex-range only with explicit range
If no explicit range is specified for an ex command, the command
itself is responsible for determining its working range. Usual evil
operators do this by the :motion property.

But if an explicit range is specified, then the region is modified to
match the region. This allows call use usual Emacs commands, which
work on the region, as ex commands. This is particularly useful if the
ex region is the visual region '<,'>.

This fixes issue #221.