Make `evil-delete-backward-word' compatible with Vim's C-w in insert state
`evil-delete-backward-word' had been implemented as "db". But this is
different from C-w if point is placed at the beginning of a line. In
this case C-w should only delete the preceding newline. This patch
reimplements `evil-delete-backward-word' as a command (instead of an
operator), which handles this special case correctly.

--HG--
extra : rebase_source : 115d4e9be9da275d76756dbe59a23b79a420dbae