Fix completion info shown by `evil-ex-update` for Emacs commands.
This is simply a more accurate interpretation of the result of the
call to `all-completions`. The current text is incomplete if there is
more than one possible completion and it is unknown if there is no
possible completion.

--HG--
extra : amend_source : d2c95fd178f7b5a5e5864fe4f5a112012841f0f6