Handle ESC -> escape translation in `input-decode-map` (fix #264)
The old ESC -> escape translation code relied on a certain handling of
the keymap code in Emacs' core. This handling has changed in Emacs
24.3 and cause bug #264 and the related Emacs bugs 13709, 13793.

The new code handles this transition not in a high-precedence minor
mode map but directly in `input-decode-map`. An incoming ESC event is
translated to either 'escape or to the default ESC prefix map
originally bound to ESC in `input-decode-map`.