Support paste-pop after visual paste (fix #283)
In contrast to a normal paste, a visual paste deletes some text before
pasting (and puts it on the kill-ring). Because `evil-paste-pop` first
undos the last operation (i.e. the paste), the original text is
restored, so the repeated paste must be executed as 'visual paste',
too, with the original visual region restored, so that the new paste
operation will again delete the original text. Furthermore, the first
visual paste (and only the first) adds the replaced text to the kill
ring, so the paste-pop must take care of this situation and skip that
text appropriately. The information that the previous paste was the
first visual paste (and thus modified the kill-ring) is saved as 6th
value in the `evil-last-paste` variable.

--HG--
extra : amend_source : 21f6b2d84b36d4c6cca1d347391462708f4ebeb3