Disable narrowing in char arguments in operator state (fix #294)
When éa motion is read in operator state, the buffer is narrowed to
the current field so that the operation is always restricted to that
field. If a character argument is read then this restriction is
visible. This fix disables the restriction temporarily if an argument
with the <C> interactive code is used, which is the case for f,F,t and
T motions.

--HG--
extra : rebase_source : 61ad2075b6d7b31dea0391151f88e4a1b4e4d1a8