Make ; and , skip adjacent characters when repeating t or T (fix #312).
The old behaviour of ; and , after t or T was to repeat that command
literally which caused point not to move. The latest Vim version now
support a configuration option cpo-; that causes the repeat command to
skip an adjacent character. This change enables this behaviour in Evil
given that `evil-repeat-char-to-skip-next` is set to non-nil (which is
the default).

--HG--
extra : amend_source : 69991678b11e9742595855ccebfdebc6b36fcfbf