Do not mark undo/redo as motions but as non-repeatable (fix #308).
Declaring those functions as motions causes the region to be *not* set
up to match the visual selection before the command is run (because
motions do not care for the region but for the correct position of
point). However, undo/redo may be used with an active region, so it is
better to declare them as normal commands that are called with the
correct region but may not be repeated.