Do not bind to undo-tree functions explicitly.
If undo-tree is enabled, it remaps `undo` and `redo` to the
corresponding undo-tree functions. Hence `u` and `C-r` both use
undo-tree automatically if enabled. Furthermore, both keys bind to the
standard `undo` and `redo` functions (with `redo.el` or `redo+.el`) if
undo-tree is disabled. Note that this makes Evil use the correct
bindings on a per-buffer basis, i.e., if undo-tree is enabled only in
few buffers.