support repeating register commands (re #392)
Because choosing a register is a command on its own, the selecting has
been recorded as single command in the repeat ring. This change
inhibits this recording but modifies the subsequent command, so that
the original register is replayed.

--HG--
extra : amend_source : cd23d76f24ae36c0128ebd6b32ab71f92221d792