initialize parser with current state in `evil-forward-quote` (fix #494)
The call to `beginning-of-defun` might move point to an invalid point
(e.g. in `ruby-mode` in a top-level position). Because
`parse-partial-sexp` assumes that point is at the beginning of a defun
unless an explicit parser state is given, this might lead to an
invalid resulting parser state (in the case of `ruby-mode` point is
not moved at all). Therefore we initialize the parser with the current
state at `point`, which should do the correct thing anyway (because
the beginning of a defun should never be contained in a string).