do not use prefix argument for state toggle functions (fix #518)
If a non-positive argument is passed to a state toggle function like
`evil-normal-state`, the state is deactivated but no other state is
activated. Because this is not useful for interactive use, the prefix
argument is disabled. In particular, an accidentally entered prefix
argument cannot set Evil to an unusable state because it is ignored.